pyproject.toml
[tool.poetry]
name = "habr-news"
version = "0.1.0"
description = "parsing favorite news from habr.com"
authors = ["Divider"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.15"
python-dotenv = "^0.19.0"
aioschedule = "^0.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
